dh012

+++ src/main/java/ch/ethz/idsc/edelweis/ParserImages.java
@@ -5,13 +5,13 @@ import java.util.Map;
 
 import org.jfree.chart.JFreeChart;
 
-import ch.ethz.idsc.sophus.util.plot.Histogram;
-import ch.ethz.idsc.sophus.util.plot.VisualSet;
 import ch.ethz.idsc.tensor.RealScalar;
 import ch.ethz.idsc.tensor.Scalar;
 import ch.ethz.idsc.tensor.Tensor;
 import ch.ethz.idsc.tensor.Tensors;
 import ch.ethz.idsc.tensor.alg.Range;
+import ch.ethz.idsc.tensor.fig.Histogram;
+import ch.ethz.idsc.tensor.fig.VisualSet;
 import ch.ethz.idsc.tensor.red.Max;
 import ch.ethz.idsc.tensor.red.Min;
 import ch.ethz.idsc.tensor.red.Tally;
@@ -30,7 +30,7 @@ public enum ParserImages {
     Tensor values = Tensors.empty();
     Map<Tensor, Long> map = Tally.of(ceil);
     for (Tensor _x : domain)
-      values.append(RealScalar.of(map.containsKey(_x) ? map.get(_x) : 0));
+      values.append(RealScalar.of(map.containsKey(_x) ? (long) map.get(_x) : (long) 0));
     VisualSet visualSet = new VisualSet();
     visualSet.setPlotLabel(bulkParser.name() + " line count histogram");
     visualSet.setAxesLabelX("lines");
